Work on core war test

+++ VM/includes/operations.h
@@ -6,7 +6,7 @@
 /*   By: krioliin <krioliin@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/12/28 15:40:03 by krioliin       #+#    #+#                */
-/*   Updated: 2020/01/19 15:55:53 by krioliin      ########   odam.nl         */
+/*   Updated: 2020/01/24 20:15:30 by krioliin      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,7 +60,7 @@ bool		op_zjmp(t_cursor *cursor, t_vm *vm);
 
 int			get_arg_val(e_argctype arg_type, uint8_t arena[MEM_SIZE],
 			t_cursor *cursor, int *padding);
-void		write_into_memory(int32_t val_to_write, uint8_t *arena, int address);
-// void		write_into_memory(int32_t val_to_write, uint8_t *arena);
+// void		write_into_memory(int32_t val_to_write, uint8_t *arena, int address);
+void	write_into_memory(long long val_to_write, uint8_t *arena, int address);
 
 #endif
